remote_control/requirements.txt
setuptools
traitlets
adafruit-circuitpython-servokit
Jetson.GPIO
Flask
python-socketio
uvicorn
fastapi
fastapi-socketio
eventlet